+++ capstone/src/main/java/com/techelevator/view/Munchy.java
@@ -1,9 +1,11 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Munchy extends Product{
 
 
-    public Munchy(String name, double price, String slot, int inventory) {
+    public Munchy(String name, BigDecimal price, String slot, int inventory) {
         super(name, price, slot, inventory);
