+++ capstone/src/main/java/com/techelevator/view/Sandwich.java
@@ -1,8 +1,10 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Sandwich extends Product {
 
-    public Sandwich(String name, double price, String slot, int inventory) {
+    public Sandwich(String name, BigDecimal price, String slot, int inventory) {
         super(name, price, slot, inventory);
 
     }
